更新		更换账本html文件名称与js文件名称 更新		调整验证码,增加个数,减少干扰线条数 新增		账本记录html,js,后台controller,service,mapper,entity,xml 修改		调整博客列表html DOM属性id的名称 修改		移动favicon.ico网站图标文件位置 修改		调整上方菜单栏数据 修改		增加主页动画(借鉴网络资源) 新增		增加redis jar包,jedis-2.9.0.jar 修改		调整RST提示语类 修改		调整user类,增加注册的属性确认密码 修改		调整usermapper.xml内的错误属性名称 修改		优化userservice注册逻辑 修改		优化util.js的注释 新增		新增账本的controller的requestMapping url{修改,删除} 新增		新增账本的删除,修改service

+++ src/main/java/com/sujianan/bean/util/Bill.java
@@ -1,5 +1,6 @@
 package com.sujianan.bean.util;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import org.springframework.format.annotation.DateTimeFormat;
@@ -13,7 +14,8 @@ public class Bill {
 
     private String reasonType;
 
-    private Double money;
+    private BigDecimal money;
+    
     @DateTimeFormat(pattern="yyyy-MM-dd")
     private Date eventTime;
 
@@ -55,11 +57,11 @@ public class Bill {
         this.reasonType = reasonType == null ? null : reasonType.trim();
     }
 
-    public Double getMoney() {
+    public BigDecimal getMoney() {
         return money;
     }
 
-    public void setMoney(Double money) {
+    public void setMoney(BigDecimal money) {
         this.money = money;
     }
